자료형 수정

+++ src/main/java/com/ridingmate/api/payload/user/dto/BikeDto.java
@@ -21,7 +21,7 @@ public class BikeDto {
     //주행거리
     private int mileage;
     //연비
-    private int fuelEfficiency;
+    private double fuelEfficiency;
     //주유 횟수
     private int countOiling;
     //정비 횟수
